arctic_landscape/obstMapCreator/requirements.txt
lxml
pycollada
numpy-stl
